style: cleaned up code a bit more and separated Factory from Dependancy Injection registrations.
