+++ NameSorter/ConsolePrompt/ConsolePromptClasses.cs
@@ -1,5 +1,5 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
-using Name_Sorter.FactoriesAndRegistrations;
+using Name_Sorter.Registrations;
 using Name_Sorter.NameClasses;
 using Name_Sorter.Sorting;
 using System;
