+++ NameSorter_NUnitTest/NameSorterTests_ConolePromptScripts.cs
@@ -2,7 +2,8 @@ using Moq;
 using Microsoft.Extensions.DependencyInjection;
 using Name_Sorter;
 using Name_Sorter.ConsolePrompt;
-using Name_Sorter.FactoriesAndRegistrations;
+using Name_Sorter.Registrations;
+using Name_Sorter.Factories;
 using Name_Sorter.Sorting;
 using System.Security.Cryptography.X509Certificates;
 using Name_Sorter.NameClasses;
