+++ NameSorter_NUnitTest/NameSorterTests_SortingClasses.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using Name_Sorter.ConsolePrompt;
-using Name_Sorter.FactoriesAndRegistrations;
+using Name_Sorter.Registrations;
+using Name_Sorter.Factories;
 using Name_Sorter.NameClasses;
 using Name_Sorter.Sorting;
 using System;
