+++ NameSorter_NUnitTest/NameSorter_NUnitTest.csproj
@@ -18,7 +18,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\NameSorter\Name Sorter.csproj" />
     <ProjectReference Include="..\NameSorter\NameSorter.csproj" />
   </ItemGroup>
